Refactors server configuration to check and print all properties

+++ services/src/main/java/io/spacedog/services/AccountResource.java
@@ -198,11 +198,10 @@ public class AccountResource extends AbstractResource {
 		if (!isTest(context))
 			Internals.get().notify(//
 					Start.get().configuration().superdogNotificationTopic(), //
-					String.format("%s got a new account", Start.get().configuration().getUrl()), //
+					String.format("%s got a new account", Start.get().configuration().url()), //
 					String.format("account backend = %s\naccount email = %s", account.backendId, account.email));
 
-		ObjectNode payloadContent = Payloads
-				.savedBuilder(true, "/v1/admin", ACCOUNT_TYPE, account.backendId, version)//
+		ObjectNode payloadContent = Payloads.savedBuilder(true, "/v1/admin", ACCOUNT_TYPE, account.backendId, version)//
 				.put(AccountResource.BACKEND_KEY, account.defaultClientKey()).build();
 
 		return Payloads.json(payloadContent, HttpStatus.CREATED)//
@@ -263,8 +262,10 @@ public class AccountResource extends AbstractResource {
 
 		ElasticHelper.get().refresh(true, ADMIN_INDEX);
 
-		FileResource.get().deleteAll();
-		ShareResource.get().deleteAll();
+		if (!isTest(context) || !Start.get().configuration().isOffline()) {
+			FileResource.get().deleteAll();
+			ShareResource.get().deleteAll();
+		}
 
 		return Payloads.success();
 	}
